- Updated symbol editor UI and made it delay loading
- Cleaned code for input format helper
